27/3-96:
Fixed the bug:
recovery was not offered on fragment groups without fragmentforms

Added Search/Replace in the Edit menu of the property editor

1/4-96:

Fixed the bug:
When opening the second codeeditor on the same fragment form,
and the ff has been changed in the first editor, the user is asked
whether to recover from the autosave file

Moved the groupFileMenu.open and performMenu.open to sif,
in this way other tools like Frigg and Freja, can insert their
own commands together with Sifs.
Take a copy these open bindings and adjust them.
This may be done differently in the future.
The same is done in siflight, compsif and usersif.

Added the following notifications to be used by Frigg:

ymersiflib.bet, fragmentBrowser:

     onAboutToCloseGroup:<
       (# fg: ^astInterface.fragmentGroup;
          OKtoClose,askSave: @boolean
            (* if OKtoClose the groupEditor is closed.
             * if askSave the user is prompted whether to save the group
             * else it is just saved (if necessary)
             *)
       enter fg[]
       do true->OKtoClose->askSave;
          INNER
       exit (OKtoClose,askSave)
       #);
     onGroupSaved:<
       (# fg: ^astInterface.fragmentGroup;
       enter fg[]
       do
          INNER
       #);
     onGroupCompiled:<
       (# fg: ^astInterface.fragmentGroup;
          OK: @boolean (* OK means no errors *)
       enter (fg[],OK)
       do
          INNER
       #);
     onGroupChecked:<
       (# fg: ^astInterface.fragmentGroup;
          OK: @boolean (* OK means no errors *)
       enter (fg[],OK)
       do
          INNER
       #);

2/4-96:

added onGroupClosed in ymersiflib (fragmentBrowser),
which is called when sif closes a groupeditor

     onGroupClosed:<
       (# fg: ^astInterface.fragmentGroup
       enter fg[]
       do INNER
       #);


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3065 678b6752-a8a0-49ff-a511-7c2a3241fcd4