Fixed the bug: Revert textediting does not reinstall the mps default error reporter. This gives problems during property editing. Solution: a ref. to mps.errorReporter is kept in edenv.mpsErrorReporter. After Revert textediting mps.errorReporter is reinstalled.
Fixed the bug: opening the property editor on an empty property list
crashes.

Fixed the bug: crashed when reopening a propertyeditor that was closed
via the System menu.

Fixed the bug: when entering an empty file name in the New dialogs the
editor crashed

Fixed the bug: when pressing a key in an empty siftexteditor the editor
crashes in onbeforechange

CVSr ----------------------------------------------------------------------


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3073 678b6752-a8a0-49ff-a511-7c2a3241fcd4