reinstalled the mps errorreporter just before parsing the properties to the prop ast
fixed the bug: enter textedit as 1. op in the editor - parse error
revert textediting - undo - crash with ref. is none in doundo

fixed the bug: when the compiler saves a group and calls groupSaved this
is not propagated to onGroupSaved

tried to fix the EOS exception in the errorreporter, but could not reproduce
the error


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3076 678b6752-a8a0-49ff-a511-7c2a3241fcd4