16/4-96:
change markAsChanged to extendMarkAsChanged in editorenvboby
(New.., New BETA prog etc.)
to bind exceptions, now it does not crash if the .ast file already
exists and is non-writable, but the "changedByOtherTool" check
gives an alert! should not happen

allow danish characters in textediting, only disallow characters < ' '

commit textediting and installed new version


18/4:

there is also a problem with closed fragment groups after compilation
if the user changes files that are NOT open in sif, because the
compiler for such files closes the fragment group and opens a
new one (see control-controller: grouplist.insertGroup),
this gives problems in the sourcebrowsers dependency graph
that still has references to the old fg.
to avoid crashes the user is now notified when any files in the DG has
been changed (whether open or not in sif) and compilation is
prohibited, but we need to solve this problem soon!!


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3090 678b6752-a8a0-49ff-a511-7c2a3241fcd4