6/5-96
Explicit close on promptforboolean in Save All, to avoid problems with
2 clicks on the OK button, will later be fixed in guienv.

Fixed the bug: when searching fragments for slots (Follow Link to Fragment)
no exceptions was bound on top.open on the body files, this caused problems
if the body file did not exist or had parse errors.

7/5-96

Making test and trace info available via menu when calling sif with -test
argument.

Fixed the bug: dump with qualification error in updatebrowser if editing
properties while having a separate editor on that group. The problem was
that the group editor of the separate editor got an propertiesChanged event
which made a call to updatebrowser, but in this case we do not have a
browser!! Now updatebrowser tests for this.

8-9/5-96

The following notification is called if the user answers no to save
the changes:

onGroupNotSaved:<
  # fg: ^astInterface.fragmentGroup enter fg[] do INNER #)


The following operation is added in order to make experiences wit
using emacs as text editor:

externalTextEdit:<
  (# textBefore,textAfter: ^text
  enter textBefore[]
  do INNER
  exit textAfter[]
  #)

When sif is called with -e as agurment, an extra menu element is
added to the Edit menu: 'External Textedit', with short cut Ctrl e.
When this command is activated the text in the current selection is
used as textBefore.  externalTextEdit is supposed to block sif
until the editing in emacs is finished, and textAfter should then contain
the new text.

As a side effect a texteditor update problem has been removed.
When entering textediting on a sublist which includes the last element
of the list the text was not updated correctly. Similar after
revert textediting. Preliminary workaround: prettyprint some more :-)


The test for donecheck is removed when following semantic links.
This means that it is now possible to follow semantic links in
files that are not semantically correct.


Fixed the bug: when using (Follow Link to SLOT)
no exceptions was bound on top.open on the origin files, this caused problems
if the origin file did not exist or had parse errors.


Changes on the fragment group level is now only reported to Freja when the
group is open in Freja.


Fixed the bug: when opening a group file from freja, an error recovery
dialog appeared, which should not appear.


committed and installed v5.0.2(6)


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3139 678b6752-a8a0-49ff-a511-7c2a3241fcd4