Fixed the bug. If the user tries to "out comment" some code by putting a (* in the start of the texteditarea and a *) in the end, the comment is lossed. For the time being a workaround is reported to the user: Textedit the enclosing block. (textediting.bet)
Added the rest of the external texteditpart, i.e. the start of am external
texteditor (specified by $EDITOR) on a temporary file placed in /temp.
The name of the file is siftext$USER$TIME.bet.
Has been tested with emacs and vi.
NB! only Unix.

emacs:

EDITOR must be set to e.g.
/usr/local/bin/emacsclient

and the following must be added to the .emacs file:

;;; sif

(server-start)

(defun my-server-edit ()
  (interactive)
  (server-edit)
  (server-edit))

; shourt cut for finishing the external text edit
(define-key global-map [?\C-e] 'my-server-edit)

(setq server-temp-file-regexp "/tmp/sif*")

Currently this feature is only available if sif is started with -e.
The external editing is activated by the Edit:External Textedit command (Ctrl-e)

We need a way to regret changes in emacs.

vi:

EDITOR must be set to e.g.
/usr/ucb/vi


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@3164 678b6752-a8a0-49ff-a511-7c2a3241fcd4