Fixed BUG in PImport: refhandler was not called on references in refeference repetitions (RefRepPTValue)
Extended refhandler to deal with references internal to the object.
Useful as e.g. origin references often refer to the same "object", and
needed to get origins of static partobjects correct, as the surrounding
object is not in any table at that point in time.

Added updateOtherReferences to referenceTable. I'm not 100% sure whether
it is really needed or if some other code guarantees that all oter
proxies referring to parts of the same object are updated when an object
is loaded?

Betarun version number: 51


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@12890 678b6752-a8a0-49ff-a511-7c2a3241fcd4